Object Exercises
Commented out the script link in the index.html so that the object exercises are the only thing outputting to the webpage.

Added four exercises that test my ability to instantiate, call, and manipulate objects in JavaScript.

The programs are as follows:

Personal Contact Book (allows a user to add, view, and search for contacts by name)

Student Grade Management (calculates a class average and populates an array of objects that include students that made below the class average)

Movie Collection Tracker (allows a user to manage a movie collection by adding and rating a movie collection and viewing the movie collection.)

Library Book Tracker (allows the user to sort an inventory of books by whether they are available or borrowed.)